# RecommendationService/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PetCare Recommendation Service API
  version: 1.0.0
servers:
  - url: http://localhost:8083/recommendation/api/v1
  - url: http://recommendation-service:8080/recommendation/api/v1
paths:
...
  /users/{userId}/pets:
    post: # Существующий
      tags: [ Pet ]
      summary: Создать питомца для пользователя
      operationId: createPetForUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPetCreateRequestDto'
      responses:
        '201':
          description: Питомец создан
          headers:
            Location:
              description: URI нового питомца
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPetResponseDto"
        '400':
          $ref: '#/components/responses/BadRequest'
        # '404': # Если проверяем userId
        #   $ref: '#/components/responses/NotFound'
    get: # Новый GET List
      tags: [ Pet ]
      summary: Получить список питомцев пользователя
      operationId: getPetsForUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список питомцев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPetResponseDto"
        # '404': # Если проверяем userId
        #   $ref: '#/components/responses/NotFound'
  /users/{userId}/pets/{petId}:
    get: # Новый GET Single
      tags: [ Pet ]
      summary: Получить питомца по ID
      operationId: getPetByIdForUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: petId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Питомец найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPetResponseDto'
        '404':
          $ref: '#/components/responses/NotFound' # Pet или User не найден
    patch: # Новый PATCH
      tags: [ Pet ]
      summary: Обновить информацию о питомце
      operationId: updatePetForUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: petId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPetUpdateRequestDto'
      responses:
        '200':
          description: Питомец обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPetResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound' # Pet или User не найден

  # --- Care Tips Recommendations (Read-Only for Pet) ---
...
components:
  schemas:
...
    UserPetResponseDto:
      type: object
      properties:
        petId:
          type: string
          format: uuid
          readOnly: true
        ownerId:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        birthDate:
          type: string
          format: date
          nullable: true
        speciesName:
          type: string
        breedName:
          type: string
          nullable: true
        allergyNames:
          type: array
          items:
            type: string # Просто массив строк
          description: Список названий аллергий

    # --- Enums ---
...
    UserPetCreateRequestDto:
      type: object
      required:
        - name
        - speciesName
      properties:
        name:
          type: string
          maxLength: 100
        birthDate:
          type: string
          format: date
          nullable: true
          description: Дата рождения должна быть сегодня или в прошлом
        speciesName:
          type: string
          maxLength: 50
        breedName:
          type: string
          maxLength: 100
          nullable: true
        allergyNames:
          type: array
          items:
            type: string # Просто массив строк
            maxLength: 100
          description: Список названий аллергий питомца
    UserPetUpdateRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          nullable: true
        birthDate:
          type: string
          format: date
          nullable: true
          description: Дата рождения должна быть сегодня или в прошлом
        speciesName:
          type: string
          maxLength: 50
          nullable: true
        breedName:
          type: string
          maxLength: 100
          nullable: true # null - не менять, "" - удалить породу
        allergyNames:
          type: array
          items:
            type: string
            maxLength: 100
          nullable: true # null - не менять, [] - очистить, [...] - установить
          description: Полный новый список названий аллергий питомца

    # --- Error Response ---
    ErrorResponse: # Переименовал для ясности, использовал стандартные поля ProblemDetail
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference [RFC3986] that identifies the problem type.
          default: "about:blank"
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
  responses:
    BadRequest:
      description: Некорректный запрос (ошибка валидации и т.п.)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...